# analysis-api/src/auth/dto/refresh-token.dto.ts
import { ApiProperty } from "@nestjs/swagger";

export class RefreshTokenDto {
  @ApiProperty({ description: "Refresh token, полученный при login или refresh" })
  refreshToken: string;
}